Double-precision camera changes
Script sync from SHA: 5250b169a2494d40f1c51d15e886b07064db6b68

+++ src/tessellatedclusters_ui.cpp
@@ -264,7 +264,7 @@ void TessellatedClusters::onUIRender()
       const glm::vec3 hitPos   = glm::unProjectZO({mousePos.x, mousePos.y, d}, view, proj, win_norm);
 
       // Set the interest position
-      glm::vec3 eye, center, up;
+      glm::dvec3 eye, center, up;
       m_info.cameraManipulator->getLookat(eye, center, up);
       m_info.cameraManipulator->setLookat(eye, hitPos, up, false);
     }
